refactored catalog client constructor
